Make char append a verified String operation
The string maturity lane had a runtime helper for appending Unicode scalars, but the public stdlib wrapper was still missing and HIR char types fell through to void in MIR declarations. Preserve char as the intended i32 ABI scalar and expose String.push_char through the owned String API.

Constraint: The parser still does not accept source-level `as` casts, so the wrapper must pass char directly instead of casting through i64.
Rejected: Keep the C bridge as long long | would rely on platform-specific widening from an i32 LLVM argument to a 64-bit C parameter.
Confidence: high
Scope-risk: narrow
Directive: Do not widen sengoo_string_push_char_status without updating the Sengoo extern signature and the i32 declaration regression together.
Tested: cargo test -p sengoo-compiler string_module_imports_owned_string_push_char -- --nocapture
Tested: cargo test -p sgc stdlib_string_push_char_appends_unicode_scalar -- --nocapture with CARGO_TARGET_DIR=target/codex-push-char
Tested: cargo test -p sgc examples_smoke_stdlib_formatting_import -- --nocapture
Tested: cargo test -p sengoo-compiler format -- --nocapture
Tested: cargo fmt --check
Tested: cargo clippy -p sengoo-compiler -p sgc --all-targets -- -D warnings
Tested: npx @fission-ai/openspec validate first-class-strings-and-formatting --strict
Tested: npx @fission-ai/openspec validate --all --strict
Not-tested: Full workspace test suite; this slice is scoped to String.push_char and char ABI lowering.

+++ tools/stdlib/runtime_string.c
@@ -401,7 +401,7 @@ long long sengoo_string_push_bool_status(long long handle, long long value) {
     return sengoo_owned_string_append_bytes(handle, text, strlen(text));
 }
 
-long long sengoo_string_push_char_status(long long handle, long long codepoint) {
+long long sengoo_string_push_char_status(long long handle, int codepoint) {
     if (codepoint < 0 || codepoint > 0x10FFFF
         || (codepoint >= 0xD800 && codepoint <= 0xDFFF)) {
         return -(long long)SENGOO_STATUS_INVALID_ARGUMENT;
